refactor: 파일경로수정, import 시 함수 이름수정
