+++ frontend/Prog-rangers-main/package.json
@@ -13,10 +13,6 @@
     "@testing-library/user-event": "^13.5.0",
     "aos": "^3.0.0-beta.6",
     "axios": "^1.4.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-react": "^7.33.2",
     "jotai": "^2.2.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -52,6 +48,7 @@
   "devDependencies": {
     "@emotion/babel-plugin": "^11.11.0",
     "@emotion/babel-preset-css-prop": "^11.11.0",
+    "eslint-plugin-react": "^7.33.2",
     "prettier": "3.0.2"
   }
 }
