+++ frontend/Prog-rangers-main/src/App.js
@@ -1,5 +1,6 @@
+// import React from 'react';
 import { Outlet } from 'react-router-dom';
-import Header from './components/Header';
+import { Header } from './components/Header';
 import './App.css';
 
 function App() {
@@ -8,7 +9,7 @@ function App() {
       <Header />
       <Outlet />
     </>
-  );
+  ); 
 }
 
 export default App;
