+++ frontend/Prog-rangers-main/src/components/Header/BeforeLoginNav.jsx
@@ -3,9 +3,9 @@ import { Link } from 'react-router-dom';
 import {
   navStyle,
   linkDefault,
-} from '../styles/headerStyle';
+} from './headerStyle';
 
-const BeforeLoginNav = () => {
+export const BeforeLoginNav = () => {
   return (
     <div className="BeforeLoginNav" css={navStyle}>
       <Link to="/problems" css={linkDefault}>
@@ -21,4 +21,3 @@ const BeforeLoginNav = () => {
   );
 };
 
-export default BeforeLoginNav;
