+++ frontend/Prog-rangers-main/src/components/Home/index.js
@@ -1,0 +1,1 @@
+export { ScrollDownArrow } from './ScrollDownArrow';
