+++ frontend/Prog-rangers-main/src/components/SignIn/index.js
@@ -1,0 +1,1 @@
+export { DefaultSignInForm } from './DefaultSignInForm';
