+++ frontend/Prog-rangers-main/src/components/SignUp/index.js
@@ -1,0 +1,2 @@
+export { SignUpFrom } from './SignUpForm';
+export { SocialLoginButtons } from './SocialLoginButtons';
