+++ frontend/Prog-rangers-main/src/pages/MainPage/MainPage.jsx
@@ -1,6 +1,6 @@
 import React, { useRef } from 'react';
 import { css } from '@emotion/react';
-import { theme } from '../styles/theme';
+import { theme } from '../../components/Header/theme';
 import { Link } from 'react-router-dom';
 import {
   secLight,
@@ -14,11 +14,11 @@ import {
   flexRowLayout,
   flexColLayout,
   buttonSytle,
-} from '../styles/MainPage';
-import ScrollDownArrow from '../components/Home/ScrollDownArrow';
-import Sec2Mockup from '../assets/main-sec2.png';
-import Sec3Mock1 from '../assets/main-sec3-1.png';
-import Sec3Mock2 from '../assets/main-sec3-2.png';
+} from './MainStyle';
+import { ScrollDownArrow } from '../../components/Home/ScrollDownArrow';
+import Sec2Mockup from '../../assets/main-sec2.png';
+import Sec3Mock1 from '../../assets/main-sec3-1.png';
+import Sec3Mock2 from '../../assets/main-sec3-2.png';
 import {
   bookPosition,
   chatPosition,
@@ -26,19 +26,19 @@ import {
   folderPosition,
   notePosition,
   penPosition,
-} from '../styles/MainPageIcons';
-import IconPen from '../assets/icons/main-pen.svg';
-import IconFolder from '../assets/icons/main-folder.svg';
-import IconNote from '../assets/icons/main-note-n-pen.svg';
-import IconCommets from '../assets/icons/main-comments.svg';
-import IconChat from '../assets/icons/main-chat.svg';
-import IconBook from '../assets/icons/main-openbook.svg';
+} from './MainPageIcons';
+import IconPen from '../../assets/icons/main-pen.svg';
+import IconFolder from '../../assets/icons/main-folder.svg';
+import IconNote from '../../assets/icons/main-note-n-pen.svg';
+import IconCommets from '../../assets/icons/main-comments.svg';
+import IconChat from '../../assets/icons/main-chat.svg';
+import IconBook from '../../assets/icons/main-openbook.svg';
 
 const displayBlock = css`
   display: block;
 `;
 
-const MainPage = () => {
+export const MainPage = () => {
   const element = useRef();
   const moveToElement = () => {
     element.current.scrollIntoView({ behavior: 'smooth' });
@@ -318,5 +318,3 @@ const MainPage = () => {
     </div>
   );
 };
-
-export default MainPage;
