+++ frontend/Prog-rangers-main/src/pages/MainPage/MainPageIcons.js
