+++ frontend/Prog-rangers-main/src/pages/MainPage/MainStyle.js
@@ -1,5 +1,5 @@
 import { css } from '@emotion/react';
-import { theme } from './theme';
+import { theme } from '../../components/Header/theme';
 
 // section
