+++ frontend/Prog-rangers-main/src/pages/MainPage/index.js
@@ -1,0 +1,1 @@
+export { MainPage } from './MainPage';
