+++ frontend/Prog-rangers-main/src/pages/NotFound.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-export default function NotFound() {
+export const NotFound = () => {
   return <div>404 Not Found</div>;
 }
