+++ frontend/Prog-rangers-main/src/pages/RegisterReview.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-export default function RegisterReview() {
+export const RegisterReview = () => {
   return <div>RegisterReview Page</div>;
 }
