+++ frontend/Prog-rangers-main/src/pages/Solutions.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-export default function Solutions() {
+export const Solutions = () => {
   return <div>Solutions page</div>;
 }
