+++ frontend/Prog-rangers-main/src/pages/index.js
@@ -1,10 +1,10 @@
-export { default as MainPage } from './MainPage';
-export { default as NotFound } from './NotFound';
-export { default as Problems } from './Problems';
-export { default as Profile } from './Profile';
-export { default as SignIn } from './SignIn';
-export { default as SignUp } from './SignUp';
-export { default as Solutions } from './Solutions';
-export { default as SolutionDetail } from './SolutionDetail';
-export { default as MyPage } from './MyPage';
-export { default as RegisterReview } from './RegisterReview';
+export { MainPage } from './MainPage';
+export { NotFound } from './NotFound';
+export { Problems } from './Problems';
+export { Profile } from './Profile';
+export { SignIn } from './SignIn';
+export { SignUp } from './SignUp';
+export { Solutions } from './Solutions';
+export { SolutionDetail } from './SolutionDetail';
+export { MyPage } from './MyPage';
+export { RegisterReview } from './RegisterReview';
